requirements.txt.txt
Flask
pandas
openpyxl
Werkzeug
google-generativeai